TEXT PRESENTATION ally methods
